refactor(fjall): delete test-only encode_event_value; relocate white-box codec tests in-crate; benches use wire::encode_frame
